+++ src/m365/aad/commands/app/app-remove.ts
@@ -17,9 +17,13 @@ export interface Options extends GlobalOptions {
   confirm?: boolean;
 }
 
-class AadAppDeleteCommand extends GraphCommand {
+class AadAppRemoveCommand extends GraphCommand {
   public get name(): string {
-    return commands.APP_DELETE;
+    return commands.APP_REMOVE;
+  }
+
+  public alias(): string[] | undefined {
+    return [commands.APP_DELETE];
   }
 
   public get description(): string {
@@ -36,6 +40,8 @@ class AadAppDeleteCommand extends GraphCommand {
   }
 
   public commandAction(logger: Logger, args: CommandArgs, cb: () => void): void {
+    this.showDeprecationWarning(logger, commands.APP_DELETE, commands.APP_REMOVE);
+
     const deleteApp: () => void = (): void => {
       this
         .getObjectId(args, logger)
@@ -65,7 +71,7 @@ class AadAppDeleteCommand extends GraphCommand {
         type: 'confirm',
         name: 'continue',
         default: false,
-        message: `Are you sure you want to delete the App?`
+        message: `Are you sure you want to remove the app?`
       }, (result: { continue: boolean }): void => {
         if (!result.continue) {
           cb();
@@ -153,4 +159,4 @@ class AadAppDeleteCommand extends GraphCommand {
   }
 }
 
-module.exports = new AadAppDeleteCommand();
+module.exports = new AadAppRemoveCommand();